split app into components and pages
