+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@astrouxds/astro-web-components": "^6.0.5",
     "core-js": "^3.6.5",
+    "moment": "^2.29.1",
     "vue": "^3.0.0"
   },
   "devDependencies": {
